memory-server/pyproject.toml
[project]
name = "memory-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastmcp>=3.1.1",
    "pydantic-settings>=2.13.1",
    "pyyaml>=6.0.3",
    "sqlalchemy>=2.0.48",
]

[tool.ruff]
target-version = "py312"
line-length = 150

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true
